@@ -1,6 +1,6 @@
 {
   "name": "bootcamp",
-  "version": "3.1.0",
+  "version": "4.0.0",
   "description": "Proyectos y tópicos del Bootcamp de @Laboratoria",
   "scripts": {
     "mdlint": "mdlint .",
@@ -34,50 +34,43 @@
     "url": "https://github.com/Laboratoria/bootcamp/issues"
   },
   "dependencies": {
-    "@material-ui/core": "^4.12.1",
+    "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.1.9",
-    "firebase": "^8.7.1",
-    "highlight.js": "^11.1.0",
+    "@testing-library/jest-dom": "^5.15.1",
+    "@testing-library/react": "^12.1.1",
+    "@testing-library/user-event": "^13.2.1",
+    "firebase": "^9.5.0",
+    "highlight.js": "^11.2.0",
     "react": "^17.0.2",
-    "react-ace": "^9.4.1",
+    "react-ace": "^9.4.4",
     "react-dom": "^17.0.2",
-    "react-intl": "^5.20.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-intl": "^5.22.0",
+    "react-router-dom": "^6.0.2",
+    "react-scripts": "^4.0.3",
     "web-vitals": "^2.1.0"
   },
   "devDependencies": {
-    "@laboratoria/curriculum-parser": "^3.0.0",
-    "@laboratoria/mdlint": "^1.2.1",
-    "@octokit/rest": "^18.6.7",
+    "@laboratoria/curriculum-parser": "^3.0.2",
+    "@laboratoria/mdlint": "^1.2.2",
+    "@octokit/rest": "^18.11.3",
     "chai": "^4.3.4",
     "chokidar": "^3.5.1",
-    "concurrently": "^6.2.0",
-    "eslint": "^7.30.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jest": "^24.3.6",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.24.0",
+    "concurrently": "^6.4.0",
+    "eslint-config-airbnb": "^19.0.1",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-jest": "^25.3.0",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-react": "^7.27.1",
     "fs-extra": "^10.0.0",
     "minimist": "^1.2.5",
     "mkdirp": "^1.0.4",
-    "mocha": "^9.0.2",
+    "mocha": "^9.1.2",
     "porch": "^1.1.4",
-    "sinon": "^11.1.1",
-    "ws": "^7.5.3"
+    "sinon": "^12.0.1",
+    "ws": "^8.3.0"
   },
   "engines": {
-    "node": ">=12.x"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+    "node": ">=14.x"
   },
   "browserslist": {
     "production": [
