@@ -7,17 +7,20 @@
   "scripts": {
     "pretest": "eslint .",
     "test": "jest --coverage",
-    "start": "serve src/",
+    "start": "serve -s src/",
     "deploy": "gh-pages -d src"
   },
   "devDependencies": {
     "@babel/core": "^7.14.3",
     "@babel/plugin-transform-modules-commonjs": "^7.14.0",
     "babel-jest": "^27.0.1",
-    "eslint": "^7.27.0",
+    "eslint": "^8.3.0",
     "gh-pages": "^3.2.0",
     "jest": "^27.0.1",
-    "serve": "^12.0.0"
+    "serve": "^13.0.2"
+  },
+  "engines": {
+    "node": ">=14.x"
   },
   "jest": {
     "testEnvironment": "jsdom"
