@@ -1,17 +1,9 @@
-import { lazy, Suspense, useEffect, useState } from 'react';
-import { BrowserRouter as Router, Switch, Route, Redirect } from 'react-router-dom';
-import IntlProvider from '../../intl/IntlProvider';
+import { useEffect, useState } from 'react';
+import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
 import { useApp } from '../../lib/app';
 import Loading from '../Loading';
+import RoutesWithIntl from './RoutesWithIntl';
 import './App.css';
-
-// Dynamic lazy imports for code splitting based on routes.
-const Home = lazy(() => import('../Home'));
-const Topic = lazy(() => import('../Topic'));
-const Project = lazy(() => import('../Project'));
-const SignIn = lazy(() => import('../SignIn'));
-const TopBar = lazy(() => import('../TopBar'));
-const Track = lazy(() => import('../Track'));
 
 const App = () => {
   const { auth } = useApp();
@@ -37,57 +29,23 @@ const App = () => {
     return <Loading />;
   }
 
+  const defaultLang = defaulLocale.split('-')[0];
+
   return (
-    <Suspense fallback={<Loading />}>
-      <Router>
-        <div className="App">
-          <Switch>
-            <Route
-              path="/:lang"
-              render={(props) => {
-                const { path, params } = props.match;
-
-                if (!['es', 'pt'].includes(params.lang)) {
-                  return (
-                    <Redirect
-                      to={`/${defaulLocale.split('-')[0]}${props.history.location.pathname}`}
-                    />
-                  );
-                }
-
-                return (
-                  <IntlProvider lang={params.lang}>
-                    <TopBar />
-                    <Switch>
-                      <Route path={`${path}/signin`}>
-                        <SignIn />
-                      </Route>
-                      <Route path={`${path}/projects/:slug`}>
-                        <Project />
-                      </Route>
-                      <Route path={`${path}/topics/:slug`}>
-                        <Topic />
-                      </Route>
-                      <Route path={`${path}/:track`}>
-                        <Track />
-                      </Route>
-                      <Route path={path}>
-                        <Home />
-                      </Route>
-                    </Switch>
-                  </IntlProvider>
-                );
-              }}
-            />
-            <Route
-              exact
-              path="/"
-              render={() => <Redirect to={`/${defaulLocale.split('-')[0]}`} />}
-            />
-          </Switch>
-        </div>
-      </Router>
-    </Suspense>
+    <Router>
+      <div className="App">
+        <Routes>
+          <Route
+            path="/"
+            element={<Navigate to={`/${defaultLang}`} />}
+          />
+          <Route
+            path="/:lang/*"
+            element={<RoutesWithIntl defaultLang={defaultLang} />}
+          />
+        </Routes>
+      </div>
+    </Router>
   );
 };
 
