@@ -1,6 +1,6 @@
 const {
   getUsers,
-} = require('../../controller/users');
+} = require('../users');
 
 describe('getUsers', () => {
   it('should get users collection', (done) => {
