@@ -2,7 +2,8 @@ const kill = require('tree-kill');
 
 module.exports = () => new Promise((resolve) => {
   if (!global.__e2e.childProcessPid) {
-    return resolve();
+    resolve();
+    return;
   }
 
   kill(global.__e2e.childProcessPid, 'SIGKILL', resolve);
