@@ -1,3 +1,4 @@
+import { Link } from 'react-router-dom';
 import { FormattedMessage } from 'react-intl';
 import { makeStyles } from '@material-ui/core/styles';
 import Divider from '@material-ui/core/Divider';
@@ -18,7 +19,7 @@ const useStyles = makeStyles(() => ({
   },
 }));
 
-const DrawerMenu = ({ drawerIsOpen, setDrawerIsOpen, history, lang }) => {
+const DrawerMenu = ({ drawerIsOpen, setDrawerIsOpen, lang }) => {
   const classes = useStyles();
   return (
     <Drawer
@@ -28,18 +29,18 @@ const DrawerMenu = ({ drawerIsOpen, setDrawerIsOpen, history, lang }) => {
     >
       <div className={classes.list}>
         <List>
-          <ListItem button onClick={() => history.push(`/${lang}`)}>
+          <ListItem button component={Link} to={`/${lang}`}>
             <ListItemIcon><HomeIcon /></ListItemIcon>
             <ListItemText primary={<FormattedMessage id="home" />} />
           </ListItem>
         </List>
         <Divider />
         <List>
-          <ListItem button onClick={() => history.push(`/${lang}/js`)}>
+          <ListItem button component={Link} to={`/${lang}/js`}>
             <ListItemIcon><DeveloperModeIcon /></ListItemIcon>
             <ListItemText primary={<FormattedMessage id="webDev" />} />
           </ListItem>
-          <ListItem button onClick={() => history.push(`/${lang}/ux`)}>
+          <ListItem button component={Link} to={`/${lang}/ux`}>
             <ListItemIcon><PeopleIcon /></ListItemIcon>
             <ListItemText primary={<FormattedMessage id="ux" />} />
           </ListItem>
